Move auxiliary commands into CommandHelpers
Their is now a BaseCommmand, BashCommand and ConfigureCommand that
are located within the general code source. However all three of these
classes, whilst not commands themselves, directly influence the command
classes. This is either done via inheritance or acting as an adaptor
(e.g. the BashCommand).

This commit moves them all into a CommandHelpers module to consolidate
the code into one location.
