+++ src/commands.rb
@@ -32,4 +32,4 @@ require 'commands/each'
 require 'commands/configure/domain'
 require 'commands/configure/group'
 require 'commands/configure/node'
-require 'bash_command'
+require 'command_helpers/bash_command'
