+++ src/commands/configure/group.rb
@@ -1,5 +1,5 @@
 
-require 'configure_command'
+require 'command_helpers/configure_command'
 require 'constants'
 
 
@@ -7,7 +7,7 @@ module Metalware
   module Commands
     module Configure
 
-      class Group < ConfigureCommand
+      class Group < CommandHelpers::ConfigureCommand
         def setup(args, _options)
           @group_name = args.first
         end
